database/connection.py
"""数据库连接管理模块"""
from typing import Optional
import config

# 伪代码：实际使用时需要根据数据库类型导入相应的驱动
# import pymysql  # MySQL
# import psycopg2  # PostgreSQL


class DatabaseConnection:
    """数据库连接管理类（伪代码实现）"""
    
    def __init__(self):
        self.connection = None
        self.config = config.DATABASE_CONFIG
    
    def connect(self):
        """建立数据库连接（伪代码）"""
        """
        实际实现示例（MySQL/PostgreSQL连接）：
        
        if self.config["type"] == "mysql":
            import pymysql
            self.connection = pymysql.connect(
                host=self.config["host"],
                port=self.config["port"],
                user=self.config["user"],
                password=self.config["password"],
                database=self.config["database"],
                charset=self.config["charset"]
            )
        elif self.config["type"] == "postgresql":
            import psycopg2
            self.connection = psycopg2.connect(
                host=self.config["host"],
                port=self.config["port"],
                user=self.config["user"],
                password=self.config["password"],
                database=self.config["database"]
            )
        """
        # 伪代码：实际使用时取消上面的注释并实现数据库连接
        pass
    
    def execute_query(self, query: str, params: Optional[dict] = None):
        """执行SQL查询（伪代码）"""
        """
        实际实现示例（MySQL/PostgreSQL查询执行）：
        
        if not self.connection:
            self.connect()
        
        cursor = self.connection.cursor()
        if params:
            cursor.execute(query, params)
        else:
            cursor.execute(query)
        
        if query.strip().upper().startswith("SELECT"):
            results = cursor.fetchall()
            columns = [desc[0] for desc in cursor.description]
            return [dict(zip(columns, row)) for row in results]
        else:
            self.connection.commit()
            return cursor.rowcount
        """
        # 伪代码：实际使用时取消上面的注释并实现数据库查询
        pass
    
    def close(self):
        """关闭数据库连接（伪代码）"""
        """
        实际实现示例（关闭数据库连接）：
        
        if self.connection:
            self.connection.close()
            self.connection = None
        """
        # 伪代码：实际使用时取消上面的注释并实现数据库连接关闭
        pass
    
    def __enter__(self):
        """上下文管理器入口"""
        self.connect()
        return self
    
    def __exit__(self, exc_type, exc_val, exc_tb):
        """上下文管理器出口"""
        self.close()


# 全局数据库连接实例
db_connection = DatabaseConnection()

