requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.9

